Added Face Tone Generation
I can now simulate the creation of tones for each face position on
the Rubik's Cube. FaceMapping.py has an example of how to send
face turns to the module for sound generation.

Next steps are to hear that sound, and decode the face turns that
occurred. That's the purpose of the new SoundReader.py.