Count the strongest of competing detected states
Sometimes, a single face is detected to be in multiple positions
simultaneously (e.g. U0, and U1, both appear above the threshold).
When this happened, the algorithm would discard the state information
for that timestep.

This commit fixes that so only the strongest of competing states is
counted. This has resulted in a large increase in the number of states
that get detected each time step.