Improve resistance to slight tone shifts
Previously, any single state tone that fell within range of
being detected as a different state was reported as a full
face turn. Because frequencies often vary slightly during the
execution of the face turn this often resulted in up to four
or five face turns being reported when only one had happened.

This commit requires a state to be continuously detected over
at least 5 time steps before it can be counted as an actual
rotation. This "5" steps was chosen because it successfully
eliminated all false positive turns.
Testing with "3" steps also yielded a successful result, but
we aren't as confident that it would resist noise as well.