Complete `Influence of Cube Noisiness` Section
After reviewing the data originally analyzed 7-8 months ago, I found
additional nuances in the differences between 2 TPS and 5 TPS data that
needed to be analyzed and explained for a complete and accurate
discussion.

This commit performs another minor reorganization of the essay
structure to focus the disussion on one metric at a time as well as
adding in new charts showing the additional nuances. The chart sizes
are also standardized to allow for better layout in the final LaTeX
document.

Finally, I wrote up the full analysis for the first metric `Influence
of Cube Noisiness` and got about 50% of the way through the second
metric `Influence of Standard Deviation Threshold`