Organize Thesis Documents and Supporting Research
The contents of this repository were not very organized, having been
cobbled together over the two years that I worked on this project. This
commit seeks to provide an extra sense of organization so that others
who visit this project have an easier time finding the information they
deem important to their own research and efforts.